FastAPI/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
requests>=2.32.0
volcengine>=1.0.0
tencentcloud-sdk-python>=3.0.1200
python-socketio[asyncio_client]>=5.11.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
